Extracted attack handling scripting
