+++ Assets/Scripts/AttackHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3e1d43bbe02b7a94e8cc4510cb270d90
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
